# feat: optimize collection content retrieval with bulk queries