## Issue

Closes #396 

## Testing

Tested on localhost. Observing ~3-5s response times on average. Prod deployments may be a tad slower when the SPA is not running on the same machine as the backend/DB.

![image](https://github.com/user-attachments/assets/ed3f2481-7d1c-4bf5-a62c-64ecf777cd8d)

Tested different batch sizes for the bulk Mongoose query. Not seeing drastically different performance:

- batch 1000 --> 5.71s
- batch 2000 --> 5.43s
- batch 5000 --> 5.84s

## Summary of changes

Replace O(n) individual database queries with O(1) bulk retrieval when fetching `x_mitre_contents` for collections. This reduces response time from ~60s to <5s for large collections.

- Add `findManyByIdAndModified()` to `BaseRepository` for bulk queries
- Add `getBulkByIdAndModified()` to `BaseService` and `AttackObjectsService`
- Override in `AttackObjectsService` to handle relationships separately
- Refactor `CollectionsService.getContents()` to use bulk retrieval

Maintains original ordering and error handling behavior while significantly improving performance for the `GET /collections/:stixId/modified/:modified` endpoint when `retrieveContents=true`.